# docs/swagger.yaml
definitions:
  entitySkill.Skill:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample swagger for Short Hack
  termsOfService: http://swagger.io/terms/
  title: Short Hack
  version: "1.0"
paths:
  /skill:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request of Creating Skill Object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entitySkill.Skill'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Adding Skill
      tags:
      - Skills
  /skill/{id}:
    get:
      consumes:
      - application/json
      description: Getting Skill by Id in detail
      parameters:
      - description: id of Skill
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Getting Skill by Id
      tags:
      - Skills
  /skills:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Getting all Skill
      tags:
      - Skills
swagger: "2.0"